+++ cmd/evml/main.go
@@ -11,6 +11,7 @@ func main() {
 	rootCmd.AddCommand(
 		cmd.NewSoloCmd(),
 		cmd.NewBabbleCmd(),
+		cmd.NewRaftCmd(),
 		cmd.VersionCmd)
 
 	if err := rootCmd.Execute(); err != nil {