+++ CHANGELOG.md
@@ -7,6 +7,7 @@ FEATURES:
 - consensus/solo: consensus implmentation that simply relays transactions from
   service to state.
 - consensus/babble: consensus implementation that uses an inmemory Babble node.
+- consensus/raft: consensus implementation using hashicorp/raft
 - engine: agent that coordinates State, Service and Consensus.
 - cmd: CLI
 - deploy: scripts to deploy local docker network